requirements.txt
alembic==1.4.3
aniso8601==7.0.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==1.16.0
eventlet==0.30.0
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-Migrate==2.5.3
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.4
gevent==20.12.1
gevent-websocket==0.10.1
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.2.1
graphql-relay==2.0.1
graphql-server-core==1.1.1
graphql-ws==0.3.1
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
Mako==1.1.4
MarkupSafe==1.1.1
promise==2.3
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.4
redis==3.5.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
